Add BooksAgentHelper to restructure BooksAgent
