+++ portfolio/src/test/java/com/google/sps/TestHelper.java
@@ -22,7 +22,6 @@ import com.google.cloud.dialogflow.v2.SessionsClient;
 import com.google.gson.Gson;
 import com.google.protobuf.InvalidProtocolBufferException;
 import com.google.protobuf.Value;
-import com.google.protobuf.util.JsonFormat;
 import com.google.sps.data.*;
 import com.google.sps.utils.*;
 import java.io.*;
@@ -309,6 +308,15 @@ public class TestHelper {
   }
 
   /**
+   * Clears all stored book information for sessionID from custom datastore.
+   *
+   * @param sessionID unique id of session to delete stored information from
+   */
+  public void deleteFromCustomDatabase(String sessionID) {
+    BooksMemoryUtils.deleteAllStoredBookInformation(sessionID, customDatastore);
+  }
+
+  /**
    * Retrieves a list of entity objects for the default user from the given query for testing
    * purposes to ensure that result in datastore match the expected.
    *
@@ -345,7 +353,6 @@ public class TestHelper {
     return fetchDatastoreEntities(category, filter);
   }
 
-
   /**
    * Retrieves a list of entity objects from the given query for testing purposes to ensure that
    * result in datastore match the expected.
@@ -360,35 +367,28 @@ public class TestHelper {
     return customDatastore.prepare(query).asList(FetchOptions.Builder.withDefaults());
   }
 
-  /**
-  * Helper method for list database verification.
-  */
-  public void checkDatabaseItems(
-      int size, String listName, List<String> expectedItems) {
+  /** Helper method for list database verification. */
+  public void checkDatabaseItems(int size, String listName, List<String> expectedItems) {
     checkDatabaseItems(size, "1", listName, expectedItems);
   }
 
-  /**
-  * Helper method for list database verification.
-  */
+  /** Helper method for list database verification. */
   public void checkDatabaseItems(
       int size, String listName, List<String> expectedItems, boolean isEmpty) {
     checkDatabaseItems(size, "1", listName, expectedItems);
   }
 
   /**
-  * Methods for database verification. Checks that database entries are equal to expected.
-  *
-  * @param size Number of items expected in user's list
-  * @param userID Current user's ID
-  * @param listName Name of the list being checked
-  * @param expectedItems List of strings containing the name of all items expected to be in the user's list database
-  */
+   * Methods for database verification. Checks that database entries are equal to expected.
+   *
+   * @param size Number of items expected in user's list
+   * @param userID Current user's ID
+   * @param listName Name of the list being checked
+   * @param expectedItems List of strings containing the name of all items expected to be in the
+   *     user's list database
+   */
   public void checkDatabaseItems(
-      int size,
-      String userID,
-      String listName,
-      List<String> expectedItems) {
+      int size, String userID, String listName, List<String> expectedItems) {
     List<Entity> databaseQuery = fetchDatastoreEntities("List", userID);
     assertEquals(size, databaseQuery.size());
     Entity entity = databaseQuery.get(0);
@@ -403,29 +403,25 @@ public class TestHelper {
     }
   }
 
-  /**
-  * Helper method for aggregate list database verification.
-  */
+  /** Helper method for aggregate list database verification. */
   public void checkAggregate(
-      String fetchName,
-      List<String> expectedItems,
-      List<Integer> expectedCounts) {
+      String fetchName, List<String> expectedItems, List<Integer> expectedCounts) {
     checkAggregate(fetchName, "1", expectedItems, expectedCounts);
   }
 
   /**
-  * Methods for aggregate database verification. Checks that database entries are equal to expected.
-  *
-  * @param fetchName Category name used to fetch subcategory from datastore
-  * @param userID Current user's ID
-  * @param expectedItems List of strings containing the name of all items expected to be in the user's list database
-  * @param expectedCount List of integers containing expected counts for each expected item of the corresponding index.
-  */
+   * Methods for aggregate database verification. Checks that database entries are equal to
+   * expected.
+   *
+   * @param fetchName Category name used to fetch subcategory from datastore
+   * @param userID Current user's ID
+   * @param expectedItems List of strings containing the name of all items expected to be in the
+   *     user's list database
+   * @param expectedCount List of integers containing expected counts for each expected item of the
+   *     corresponding index.
+   */
   public void checkAggregate(
-      String fetchName,
-      String userID,
-      List<String> expectedItems,
-      List<Integer> expectedCounts) {
+      String fetchName, String userID, List<String> expectedItems, List<Integer> expectedCounts) {
     Filter filter =
         new CompositeFilter(
             CompositeFilterOperator.AND,
@@ -442,29 +438,25 @@ public class TestHelper {
     }
   }
 
-  /**
-  * Helper method for fractional aggregate list database verification.
-  */
+  /** Helper method for fractional aggregate list database verification. */
   public void checkFracAggregate(
-      String fetchName,
-      List<String> expectedItems,
-      List<Double> expectedCounts) {
+      String fetchName, List<String> expectedItems, List<Double> expectedCounts) {
     checkFracAggregate(fetchName, "1", expectedItems, expectedCounts);
   }
 
   /**
-  * Methods for aggregate database verification. Checks that database entries are equal to expected.
-  *
-  * @param fetchName Category name used to fetch subcategory from datastore
-  * @param userID Current user's ID
-  * @param expectedItems List of strings containing the name of all items expected to be in the user's list database
-  * @param expectedCount List of doubles containing expected fractional values for each expected item of the corresponding index.
-  */
+   * Methods for aggregate database verification. Checks that database entries are equal to
+   * expected.
+   *
+   * @param fetchName Category name used to fetch subcategory from datastore
+   * @param userID Current user's ID
+   * @param expectedItems List of strings containing the name of all items expected to be in the
+   *     user's list database
+   * @param expectedCount List of doubles containing expected fractional values for each expected
+   *     item of the corresponding index.
+   */
   public void checkFracAggregate(
-      String fetchName,
-      String userID,
-      List<String> expectedItems,
-      List<Double> expectedCounts) {
+      String fetchName, String userID, List<String> expectedItems, List<Double> expectedCounts) {
     Filter filter =
         new CompositeFilter(
             CompositeFilterOperator.AND,
@@ -483,16 +475,16 @@ public class TestHelper {
     }
   }
 
-/**
-  * Populates a user list database with the given items and frequencies out of the total number of 
-  * lists created.
-  *
-  * @param userID Current user's ID
-  * @param size Number of lists of the same name created by the user
-  * @param items List of pairs of strings containing the name of all items expected and integer containing the number of times added to past grocery lists.
-  */
-  public void makeUserList(
-       String userID, int size, List<Pair<String, Integer>> items)
+  /**
+   * Populates a user list database with the given items and frequencies out of the total number of
+   * lists created.
+   *
+   * @param userID Current user's ID
+   * @param size Number of lists of the same name created by the user
+   * @param items List of pairs of strings containing the name of all items expected and integer
+   *     containing the number of times added to past grocery lists.
+   */
+  public void makeUserList(String userID, int size, List<Pair<String, Integer>> items)
       throws InvalidProtocolBufferException, IOException {
     setUser("test@example.com", userID);
     for (int i = 0; i < size; i++) {
